requirements.txt
requests>=2.31.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
fake-useragent>=1.4.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.8.0
html2text>=2020.1.16
crawl4ai>=0.3.0
playwright>=1.40.0
python-dateutil>=2.8.0